third_party/blink/web_tests/platform/linux-chrome/external/wpt/html-aam/roles-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] el-form
  assert_equals: <form data-testname="el-form" data-expectedrole="form" class="ex"><input></form> expected "form" but got "Section"
Harness: the test ran to completion.
